+++ src/XactCopy.Core/Models/OverwritePolicy.vb
@@ -3,5 +3,6 @@ Namespace Models
         Overwrite = 0
         SkipExisting = 1
         OverwriteIfSourceNewer = 2
+        Ask = 3
     End Enum
 End Namespace